merges the upstream master, updates readme
